# docs/core/openapi.json
{
  "components": {
    "schemas": {
...
      "GenerateSmartReplyRequest": {
        "description": "Request to generate Smart Reply preview.",
        "properties": {
          "field_overrides": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional manual overrides for specific fields",
            "title": "Field Overrides"
          },
          "image_id": {
            "description": "Received image to reply to (public UUID)",
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          },
          "template_id": {
            "default": "qsl_card",
            "description": "Template to use",
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "image_id"
        ],
        "title": "GenerateSmartReplyRequest",
        "type": "object"
      },
      "GenerateSmartReplyResponse": {
        "description": "Response from Smart Reply generation.",
        "properties": {
          "estimated_tx_duration": {
            "description": "Estimated transmit duration in seconds",
            "title": "Estimated Tx Duration",
            "type": "integer"
          },
          "preview_id": {
            "description": "Preview identifier for transmission",
            "format": "uuid",
            "title": "Preview Id",
            "type": "string"
          },
          "preview_image_path": {
            "description": "Path to preview image",
            "title": "Preview Image Path",
            "type": "string"
          },
          "template_data": {
            "additionalProperties": true,
            "description": "All field values used",
            "title": "Template Data",
            "type": "object"
          }
        },
        "required": [
          "preview_id",
          "preview_image_path",
          "template_data",
          "estimated_tx_duration"
        ],
        "title": "GenerateSmartReplyResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TransmitSmartReplyRequest": {
        "description": "Request to transmit Smart Reply.",
        "properties": {
          "device_id": {
            "description": "Audio output device ID",
            "title": "Device Id",
            "type": "string"
          },
          "mode": {
            "description": "SSTV mode (ScottieS1, MartinM1, Robot36)",
            "title": "Mode",
            "type": "string"
          },
          "ptt_method": {
            "description": "PTT method (none, serial, vox)",
            "title": "Ptt Method",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "device_id",
          "ptt_method"
        ],
        "title": "TransmitSmartReplyRequest",
        "type": "object"
      },
      "TransmitSmartReplyResponse": {
        "description": "Response from Smart Reply transmission.",
        "properties": {
          "status": {
            "description": "Transmission status",
            "title": "Status",
            "type": "string"
          },
          "tx_id": {
            "description": "Transmit session ID",
            "format": "uuid",
            "title": "Tx Id",
            "type": "string"
          }
        },
        "required": [
          "tx_id",
          "status"
        ],
        "title": "TransmitSmartReplyResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "SSTV (Slow-Scan Television) encode/decode engine",
    "title": "SSTeVe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/smart_reply/generate": {
      "post": {
        "description": "Generate Smart Reply preview with auto-populated fields.\n\nAuto-populates template fields from image metadata using fallback hierarchy:\n1. User override (manual entry)\n2. Image metadata (callsign, frequency, SNR from decode)\n3. Configuration defaults (operator callsign)\n4. Placeholder text (\"N/A\", \"Unknown\")\n\nArgs:\n    request: Generation request with image_id, template_id, and optional overrides\n    db: Database session\n    template_engine: Template engine instance\n\nReturns:\n    Preview information with path and populated fields\n\nRaises:\n    HTTPException: If image not found, template missing, or callsign required",
        "operationId": "generate_smart_reply_api_v1_smart_reply_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSmartReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateSmartReplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Smart Reply",
        "tags": [
          "Smart Reply",
          "smart_reply"
        ]
      }
    },
...
    "/api/v1/smart_reply/transmit/{preview_id}": {
      "post": {
        "description": "Transmit Smart Reply preview image.\n\nArgs:\n    preview_id: Preview ID from generate endpoint\n    request: Transmission parameters (mode, device, PTT)\n    db: Database session\n\nReturns:\n    Transmission session information\n\nRaises:\n    HTTPException: If preview not found or transmission fails",
        "operationId": "transmit_smart_reply_api_v1_smart_reply_transmit__preview_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "preview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preview Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransmitSmartReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransmitSmartReplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transmit Smart Reply",
        "tags": [
          "Smart Reply",
          "smart_reply"
        ]
      }
    },
...
  }
}